feat: refactor file sub to x dirs plus clippy fix
