+++ sn_cli/src/subcommands/transfer/directory/file.rs
@@ -1,0 +1,185 @@
+use crate::subcommands::transfer::application::upload::UploadedFile;
+use crate::subcommands::transfer::application::upload::UPLOADED_FILES;
+use crate::subcommands::transfer::application::{download, estimate, upload};
+use clap::Parser;
+use color_eyre::{eyre::eyre, Help, Result};
+use sn_client::{Client, FilesApi, BATCH_SIZE};
+use sn_protocol::storage::{Chunk, ChunkAddress, RetryStrategy};
+use std::ffi::OsString;
+use std::path::{Path, PathBuf};
+use xor_name::XorName;
+
+/// Options to configure different aspects of the logic to upload files
+#[derive(Clone)]
+pub struct FilesUploadOptions {
+    pub make_data_public: bool,
+    pub verify_store: bool,
+    pub batch_size: usize,
+    pub retry_strategy: RetryStrategy,
+}
+
+#[derive(Parser, Debug)]
+pub enum FilesCmds {
+    Estimate {
+        /// The location of the file(s) to upload. Can be a file or a directory.
+        #[clap(name = "path", value_name = "PATH")]
+        path: PathBuf,
+        /// Should the file be made accessible to all. (This is irreversible)
+        #[clap(long, name = "make_public", default_value = "false", short = 'p')]
+        make_data_public: bool,
+    },
+    Upload {
+        /// The location of the file(s) to upload.
+        ///
+        /// Can be a file or a directory.
+        #[clap(name = "path", value_name = "PATH")]
+        path: PathBuf,
+        /// The batch_size to split chunks into parallel handling batches
+        /// during payment and upload processing.
+        #[clap(long, default_value_t = BATCH_SIZE, short='b')]
+        batch_size: usize,
+        /// Should the file be made accessible to all. (This is irreversible)
+        #[clap(long, name = "make_public", default_value = "false", short = 'p')]
+        make_data_public: bool,
+        /// Set the strategy to use on chunk upload failure. Does not modify the spend failure retry attempts yet.
+        ///
+        /// Choose a retry strategy based on effort level, from 'quick' (least effort), through 'balanced',
+        /// to 'persistent' (most effort).
+        #[clap(long, default_value_t = RetryStrategy::Balanced, short = 'r', help = "Sets the retry strategy on upload failure. Options: 'quick' for minimal effort, 'balanced' for moderate effort, or 'persistent' for maximum effort.")]
+        retry_strategy: RetryStrategy,
+    },
+    Download {
+        /// The name to apply to the downloaded file.
+        ///
+        /// If the name argument is used, the address argument must also be supplied.
+        ///
+        /// If neither are, all the files uploaded by the current user will be downloaded again.
+        #[clap(name = "name")]
+        file_name: Option<OsString>,
+        /// The hex address of a file.
+        ///
+        /// If the address argument is used, the name argument must also be supplied.
+        ///
+        /// If neither are, all the files uploaded by the current user will be downloaded again.
+        #[clap(name = "address")]
+        file_addr: Option<String>,
+        /// Flagging whether to show the holders of the uploaded chunks.
+        /// Default to be not showing.
+        #[clap(long, name = "show_holders", default_value = "false")]
+        show_holders: bool,
+        /// The batch_size for parallel downloading
+        #[clap(long, default_value_t = BATCH_SIZE , short='b')]
+        batch_size: usize,
+        /// Set the strategy to use on downloads failure.
+        ///
+        /// Choose a retry strategy based on effort level, from 'quick' (least effort), through 'balanced',
+        /// to 'persistent' (most effort).
+        #[clap(long, default_value_t = RetryStrategy::Quick, short = 'r', help = "Sets the retry strategy on download failure. Options: 'quick' for minimal effort, 'balanced' for moderate effort, or 'persistent' for maximum effort.")]
+        retry_strategy: RetryStrategy,
+    },
+}
+
+pub(crate) async fn files_cmds(
+    cmds: FilesCmds,
+    client: &Client,
+    root_dir: &Path,
+    verify_store: bool,
+) -> Result<()> {
+    match cmds {
+        FilesCmds::Estimate {
+            path,
+            make_data_public,
+        } => estimate::estimate_cost(path, make_data_public, client, root_dir).await?,
+        FilesCmds::Upload {
+            path,
+            batch_size,
+            retry_strategy,
+            make_data_public,
+        } => {
+            upload::upload_files(
+                path,
+                client,
+                root_dir.to_path_buf(),
+                FilesUploadOptions {
+                    make_data_public,
+                    verify_store,
+                    batch_size,
+                    retry_strategy,
+                },
+            )
+            .await?
+        }
+        FilesCmds::Download {
+            file_name,
+            file_addr,
+            show_holders,
+            batch_size,
+            retry_strategy,
+        } => {
+            if (file_name.is_some() && file_addr.is_none())
+                || (file_addr.is_some() && file_name.is_none())
+            {
+                return Err(
+                    eyre!("Both the name and address must be supplied if either are used")
+                        .suggestion(
+                        "Please run the command again in the form 'files upload <name> <address>'",
+                    ),
+                );
+            }
+
+            let download_dir = dirs_next::download_dir().unwrap_or(root_dir.to_path_buf());
+            let files_api: FilesApi = FilesApi::new(client.clone(), download_dir.clone());
+
+            match (file_name, file_addr) {
+                (Some(file_name), Some(address_provided)) => {
+                    let bytes =
+                        hex::decode(&address_provided).expect("Input address is not a hex string");
+                    let xor_name_provided = XorName(
+                        bytes
+                            .try_into()
+                            .expect("Failed to parse XorName from hex string"),
+                    );
+                    // try to read the data_map if it exists locally.
+                    let uploaded_files_path = root_dir.join(UPLOADED_FILES);
+                    let expected_data_map_location = uploaded_files_path.join(address_provided);
+                    let local_data_map = {
+                        if expected_data_map_location.exists() {
+                            let uploaded_file_metadata =
+                                UploadedFile::read(&expected_data_map_location)?;
+
+                            uploaded_file_metadata.data_map.map(|bytes| Chunk {
+                                address: ChunkAddress::new(xor_name_provided),
+                                value: bytes,
+                            })
+                        } else {
+                            None
+                        }
+                    };
+
+                    download::download_file(
+                        files_api,
+                        xor_name_provided,
+                        (file_name, local_data_map),
+                        &download_dir,
+                        show_holders,
+                        batch_size,
+                        retry_strategy,
+                    )
+                    .await
+                }
+                _ => {
+                    println!("Attempting to download all files uploaded by the current user...");
+                    download::download_files(
+                        &files_api,
+                        root_dir,
+                        show_holders,
+                        batch_size,
+                        retry_strategy,
+                    )
+                    .await?
+                }
+            }
+        }
+    };
+    Ok(())
+}
