+++ sn_cli/src/subcommands/transfer/directory/folder.rs
@@ -6,12 +6,13 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::subcommands::{acc_packet::AccountPacket, files::FilesUploadOptions};
+use crate::subcommands::acc_packet::AccountPacket;
 
 use sn_client::{Client, BATCH_SIZE};
 
 use sn_protocol::storage::{RegisterAddress, RetryStrategy};
 
+use crate::subcommands::transfer::directory::file::FilesUploadOptions;
 use clap::Parser;
 use color_eyre::{eyre::eyre, Result};
 use std::{
