+++ sn_node_manager/src/config.rs
@@ -113,8 +113,8 @@ pub fn create_owned_dir(path: PathBuf, owner: &str) -> Result<()> {
 }
 
 #[cfg(windows)]
-pub fn create_owned_dir(path: PathBuf, _owner: &str) -> Result<()> {
-    std::fs::create_dir_all(&path)?;
+pub fn create_owned_dir(path_buf: PathBuf, _owner: &str) -> Result<()> {
+    std::fs::create_dir_all(path_buf.as_path())?;
     Ok(())
 }
 
@@ -122,7 +122,7 @@ pub fn create_owned_dir(path: PathBuf, _owner: &str) -> Result<()> {
 pub fn get_node_registry_path() -> Result<PathBuf> {
     let path = Path::new("C:\\ProgramData\\safenode-manager");
     if !path.exists() {
-        std::fs::create_dir_all(&path)?;
+        std::fs::create_dir_all(path)?;
     }
     Ok(path.join("node_registry.json"))
 }
